Pull weather data for me and print it to command line
This is a simple tool to pull weather data for my location and print it
to the command line.  It shows the short name of the day, the high and
low temps, the amount of rain predicted and the chance of precipitation.
Relies on a Weather Underground API key being available in WU_KEY env
var.

Plans:
  Move to clean architecture
  Add support for command line args or env vars
  Add location argument support
  Add different print formats